refactor: PR feedback 반영
