+++ .gitignore
@@ -4,7 +4,6 @@ build/
 target/
 !**/src/main/**/target/
 !**/src/test/**/target/
-src/main/resources/application.properties
 
 
 ### STS ###
