+++ src/main/java/masil/backend/global/security/config/SecurityConfig.java
@@ -6,15 +6,13 @@ import masil.backend.global.security.handler.CustomAccessDeniedHandler;
 import masil.backend.global.security.handler.CustomAuthenticationEntryPoint;
 import masil.backend.global.security.handler.OAuth2SuccessHandler;
 import masil.backend.global.security.handler.OAuth2FailureHandler;
+import masil.backend.modules.member.service.CustomOAuth2UserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
-import org.springframework.security.config.annotation.web.configurers.FormLoginConfigurer;
-import org.springframework.security.config.annotation.web.configurers.HttpBasicConfigurer;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
