+++ src/main/java/masil/backend/modules/member/dto/response/OAuth2ErrorResponse.java
@@ -1,0 +1,6 @@
+package masil.backend.modules.member.dto.response;
+
+public record OAuth2ErrorResponse(
+        String message,
+        String details
+) {}
