+++ src/main/java/masil/backend/modules/member/service/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package masil.backend.global.security.config;
+package masil.backend.modules.member.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,14 +19,6 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class CustomOAuth2UserService extends DefaultOAuth2UserService {
-
-    /**
-     * OAuth2 사용자 정보를 로드
-     * 
-     * @param userRequest OAuth2 사용자 요청 정보
-     * @return OAuth2User OAuth2 사용자 정보
-     * @throws OAuth2AuthenticationException OAuth2 인증 예외
-     */
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
         try {
