+++ src/main/resources/application.properties
@@ -1,0 +1,35 @@
+spring.application.name=backend
+
+# DB 설정 (민감 정보 제외)
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.h2.console.enabled=true
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+spring.jpa.hibernate.ddl-auto=update
+
+# OAuth2 Google 설정 (민감 정보 제외)
+spring.security.oauth2.client.registration.google.scope=openid,profile,email
+spring.security.oauth2.client.registration.google.redirect-uri=http://localhost:8080/login/oauth2/code/google
+
+# OAuth2 Google 제공자 설정 (공개 설정)
+spring.security.oauth2.client.provider.google.authorization-uri=https://accounts.google.com/o/oauth2/auth
+spring.security.oauth2.client.provider.google.token-uri=https://oauth2.googleapis.com/token
+spring.security.oauth2.client.provider.google.user-info-uri=https://www.googleapis.com/oauth2/v2/userinfo
+spring.security.oauth2.client.provider.google.user-name-attribute=sub
+
+# 세션 설정 (OAuth2 인증 요청 저장을 위해)
+server.servlet.session.timeout=30m
+server.servlet.session.cookie.name=JSESSIONID
+server.servlet.session.cookie.http-only=true
+server.servlet.session.cookie.secure=false
+server.servlet.session.cookie.same-site=lax
+
+# 디버깅을 위한 로그 설정
+logging.level.org.springframework.security=DEBUG
+logging.level.org.springframework.security.oauth2=DEBUG
+logging.level.masil.backend=DEBUG
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
+# 민감한 정보는 application-secret.properties에서 로드
+spring.profiles.include=secret
